+++ control-plane/connect-inject/webhook/consul_dataplane_sidecar.go
@@ -37,7 +37,7 @@ func (w *MeshWebhook) consulDataplaneSidecar(namespace corev1.Namespace, pod cor
 	}
 
 	multiPort := mpi.serviceName != ""
-	cmd, err := w.getContainerSidecarCommand(namespace, mpi, bearerTokenFile, pod)
+	args, err := w.getContainerSidecarArgs(namespace, mpi, bearerTokenFile, pod)
 	if err != nil {
 		return corev1.Container{}, err
 	}
@@ -82,7 +82,7 @@ func (w *MeshWebhook) consulDataplaneSidecar(namespace corev1.Namespace, pod cor
 				MountPath: "/consul/connect-inject",
 			},
 		},
-		Command:        cmd,
+		Args:           args,
 		ReadinessProbe: probe,
 		LivenessProbe:  probe,
 	}
@@ -136,7 +136,7 @@ func (w *MeshWebhook) consulDataplaneSidecar(namespace corev1.Namespace, pod cor
 	return container, nil
 }
 
-func (w *MeshWebhook) getContainerSidecarCommand(namespace corev1.Namespace, mpi multiPortInfo, bearerTokenFile string, pod corev1.Pod) ([]string, error) {
+func (w *MeshWebhook) getContainerSidecarArgs(namespace corev1.Namespace, mpi multiPortInfo, bearerTokenFile string, pod corev1.Pod) ([]string, error) {
 	proxyIDFileName := "/consul/connect-inject/proxyid"
 	if mpi.serviceName != "" {
 		proxyIDFileName = fmt.Sprintf("/consul/connect-inject/proxyid-%s", mpi.serviceName)
@@ -153,23 +153,21 @@ func (w *MeshWebhook) getContainerSidecarCommand(namespace corev1.Namespace, mpi
 		envoyConcurrency = int(val)
 	}
 
-	cmd := []string{
-		"consul-dataplane",
-		fmt.Sprintf("-addresses=%q", w.ConsulAddress),
+	args := []string{
+		"-addresses", w.ConsulAddress,
 		"-grpc-port=" + strconv.Itoa(w.ConsulConfig.GRPCPort),
-		"-proxy-service-id=" + fmt.Sprintf("$(cat %s)", proxyIDFileName),
-		"-service-node-name=${DP_SERVICE_NODE_NAME}",
+		"-proxy-service-id-path=" + proxyIDFileName,
 		"-log-level=" + w.LogLevel,
 		"-log-json=" + strconv.FormatBool(w.LogJSON),
 		"-envoy-concurrency=" + strconv.Itoa(envoyConcurrency),
 	}
 
 	if w.SkipServerWatch {
-		cmd = append(cmd, "-server-watch-disabled=true")
+		args = append(args, "-server-watch-disabled=true")
 	}
 
 	if w.AuthMethod != "" {
-		cmd = append(cmd,
+		args = append(args,
 			"-credential-type=login",
 			"-login-auth-method="+w.AuthMethod,
 			"-login-bearer-token-path="+bearerTokenFile,
@@ -177,34 +175,34 @@ func (w *MeshWebhook) getContainerSidecarCommand(namespace corev1.Namespace, mpi
 		)
 		if w.EnableNamespaces {
 			if w.EnableK8SNSMirroring {
-				cmd = append(cmd, "-login-namespace=default")
+				args = append(args, "-login-namespace=default")
 			} else {
-				cmd = append(cmd, "-login-namespace="+w.consulNamespace(namespace.Name))
+				args = append(args, "-login-namespace="+w.consulNamespace(namespace.Name))
 			}
 		}
 		if w.ConsulPartition != "" {
-			cmd = append(cmd, "-login-partition="+w.ConsulPartition)
+			args = append(args, "-login-partition="+w.ConsulPartition)
 		}
 	}
 	if w.EnableNamespaces {
-		cmd = append(cmd, "-service-namespace="+w.consulNamespace(namespace.Name))
+		args = append(args, "-service-namespace="+w.consulNamespace(namespace.Name))
 	}
 	if w.ConsulPartition != "" {
-		cmd = append(cmd, "-service-partition="+w.ConsulPartition)
+		args = append(args, "-service-partition="+w.ConsulPartition)
 	}
 	if w.TLSEnabled {
 		if w.ConsulTLSServerName != "" {
-			cmd = append(cmd, "-tls-server-name="+w.ConsulTLSServerName)
+			args = append(args, "-tls-server-name="+w.ConsulTLSServerName)
 		}
 		if w.ConsulCACert != "" {
-			cmd = append(cmd, "-ca-certs="+constants.ConsulCAFile)
+			args = append(args, "-ca-certs="+constants.ConsulCAFile)
 		}
 	} else {
-		cmd = append(cmd, "-tls-disabled")
+		args = append(args, "-tls-disabled")
 	}
 
 	if mpi.serviceName != "" {
-		cmd = append(cmd, fmt.Sprintf("-envoy-admin-bind-port=%d", 19000+mpi.serviceIndex))
+		args = append(args, fmt.Sprintf("-envoy-admin-bind-port=%d", 19000+mpi.serviceIndex))
 	}
 
 	metricsServer, err := w.MetricsConfig.ShouldRunMergedMetricsServer(pod)
@@ -217,7 +215,7 @@ func (w *MeshWebhook) getContainerSidecarCommand(namespace corev1.Namespace, mpi
 		if err != nil {
 			return nil, fmt.Errorf("unable to determine if merged metrics port: %w", err)
 		}
-		cmd = append(cmd, "-telemetry-prom-scrape-path="+prometheusScrapePath,
+		args = append(args, "-telemetry-prom-scrape-path="+prometheusScrapePath,
 			"-telemetry-prom-merge-port="+mergedMetricsPort)
 
 		serviceMetricsPath := w.MetricsConfig.ServiceMetricsPath(pod)
@@ -227,7 +225,7 @@ func (w *MeshWebhook) getContainerSidecarCommand(namespace corev1.Namespace, mpi
 		}
 
 		if serviceMetricsPath != "" && serviceMetricsPort != "" {
-			cmd = append(cmd, "-telemetry-prom-service-metrics-url="+fmt.Sprintf("http://127.0.0.1:%s%s", serviceMetricsPort, serviceMetricsPath))
+			args = append(args, "-telemetry-prom-service-metrics-url="+fmt.Sprintf("http://127.0.0.1:%s%s", serviceMetricsPort, serviceMetricsPath))
 		}
 
 		// Pull the TLS config from the relevant annotations.
@@ -263,7 +261,7 @@ func (w *MeshWebhook) getContainerSidecarCommand(namespace corev1.Namespace, mpi
 				return nil, fmt.Errorf("must set %q when providing prometheus TLS config", constants.AnnotationPrometheusKeyFile)
 			}
 			// TLS config has been validated, add them to the consul-dataplane cmd args
-			cmd = append(cmd, "-telemetry-prom-ca-certs-file="+prometheusCAFile,
+			args = append(args, "-telemetry-prom-ca-certs-file="+prometheusCAFile,
 				"-telemetry-prom-ca-certs-path="+prometheusCAPath,
 				"-telemetry-prom-cert-file="+prometheusCertFile,
 				"-telemetry-prom-key-file="+prometheusKeyFile)
@@ -273,7 +271,7 @@ func (w *MeshWebhook) getContainerSidecarCommand(namespace corev1.Namespace, mpi
 	// If Consul DNS is enabled, we want to configure consul-dataplane to be the DNS proxy
 	// for Consul DNS in the pod.
 	if w.EnableConsulDNS {
-		cmd = append(cmd, "-consul-dns-bind-port="+strconv.Itoa(consulDataplaneDNSBindPort))
+		args = append(args, "-consul-dns-bind-port="+strconv.Itoa(consulDataplaneDNSBindPort))
 	}
 
 	var envoyExtraArgs []string
@@ -307,12 +305,10 @@ func (w *MeshWebhook) getContainerSidecarCommand(namespace corev1.Namespace, mpi
 		}
 	}
 	if envoyExtraArgs != nil {
-		cmd = append(cmd, "--")
-		cmd = append(cmd, envoyExtraArgs...)
+		args = append(args, "--")
+		args = append(args, envoyExtraArgs...)
 	}
-
-	cmd = append([]string{"/bin/sh", "-ec"}, strings.Join(cmd, " "))
-	return cmd, nil
+	return args, nil
 }
 
 func (w *MeshWebhook) sidecarResources(pod corev1.Pod) (corev1.ResourceRequirements, error) {