+++ control-plane/connect-inject/webhook/consul_dataplane_sidecar_test.go
@@ -3,6 +3,7 @@ package webhook
 import (
 	"fmt"
 	"strconv"
+	"strings"
 	"testing"
 
 	"github.com/hashicorp/consul-k8s/control-plane/connect-inject/constants"
@@ -173,13 +174,10 @@ func TestHandlerConsulDataplaneSidecar(t *testing.T) {
 
 			container, err := w.consulDataplaneSidecar(testNS, pod, multiPortInfo{})
 			require.NoError(t, err)
-			expCmd := []string{
-				"/bin/sh", "-ec",
-				"consul-dataplane -addresses=\"1.1.1.1\" -grpc-port=" + strconv.Itoa(w.ConsulConfig.GRPCPort) +
-					" -proxy-service-id=$(cat /consul/connect-inject/proxyid) " +
-					"-service-node-name=${DP_SERVICE_NODE_NAME} -log-level=" + w.LogLevel + " -log-json=" + strconv.FormatBool(w.LogJSON) + " -envoy-concurrency=0" + c.additionalExpCmdArgs,
-			}
-			require.Equal(t, expCmd, container.Command)
+			expCmd := "-addresses 1.1.1.1 -grpc-port=" + strconv.Itoa(w.ConsulConfig.GRPCPort) +
+				" -proxy-service-id-path=/consul/connect-inject/proxyid " +
+				"-log-level=" + w.LogLevel + " -log-json=" + strconv.FormatBool(w.LogJSON) + " -envoy-concurrency=0" + c.additionalExpCmdArgs
+			require.Equal(t, expCmd, strings.Join(container.Args, " "))
 
 			if w.AuthMethod != "" {
 				require.Equal(t, container.VolumeMounts, []corev1.VolumeMount{
@@ -278,7 +276,7 @@ func TestHandlerConsulDataplaneSidecar_Concurrency(t *testing.T) {
 				require.EqualError(t, err, c.expErr)
 			} else {
 				require.NoError(t, err)
-				require.Contains(t, container.Command[2], c.expFlags)
+				require.Contains(t, strings.Join(container.Args, " "), c.expFlags)
 			}
 		})
 	}
@@ -301,7 +299,7 @@ func TestHandlerConsulDataplaneSidecar_DNSProxy(t *testing.T) {
 	}
 	container, err := h.consulDataplaneSidecar(testNS, pod, multiPortInfo{})
 	require.NoError(t, err)
-	require.Contains(t, container.Command[2], "-consul-dns-bind-port=8600")
+	require.Contains(t, container.Args, "-consul-dns-bind-port=8600")
 }
 
 func TestHandlerConsulDataplaneSidecar_Multiport(t *testing.T) {
@@ -366,20 +364,20 @@ func TestHandlerConsulDataplaneSidecar_Multiport(t *testing.T) {
 					serviceName:  "web-admin",
 				},
 			}
-			expCommand := [][]string{
-				{"/bin/sh", "-ec", "consul-dataplane -addresses=\"1.1.1.1\" -grpc-port=8502 -proxy-service-id=$(cat /consul/connect-inject/proxyid-web) " +
-					"-service-node-name=${DP_SERVICE_NODE_NAME} -log-level=info -log-json=false -envoy-concurrency=0 -tls-disabled -envoy-admin-bind-port=19000 -- --base-id 0"},
-				{"/bin/sh", "-ec", "consul-dataplane -addresses=\"1.1.1.1\" -grpc-port=8502 -proxy-service-id=$(cat /consul/connect-inject/proxyid-web-admin) " +
-					"-service-node-name=${DP_SERVICE_NODE_NAME} -log-level=info -log-json=false -envoy-concurrency=0 -tls-disabled -envoy-admin-bind-port=19001 -- --base-id 1"},
+			expArgs := []string{
+				"-addresses 1.1.1.1 -grpc-port=8502 -proxy-service-id-path=/consul/connect-inject/proxyid-web " +
+					"-log-level=info -log-json=false -envoy-concurrency=0 -tls-disabled -envoy-admin-bind-port=19000 -- --base-id 0",
+				"-addresses 1.1.1.1 -grpc-port=8502 -proxy-service-id-path=/consul/connect-inject/proxyid-web-admin " +
+					"-log-level=info -log-json=false -envoy-concurrency=0 -tls-disabled -envoy-admin-bind-port=19001 -- --base-id 1",
 			}
 			if aclsEnabled {
-				expCommand = [][]string{
-					{"/bin/sh", "-ec", "consul-dataplane -addresses=\"1.1.1.1\" -grpc-port=8502 -proxy-service-id=$(cat /consul/connect-inject/proxyid-web) " +
-						"-service-node-name=${DP_SERVICE_NODE_NAME} -log-level=info -log-json=false -envoy-concurrency=0 -credential-type=login -login-auth-method=test-auth-method " +
-						"-login-bearer-token-path=/var/run/secrets/kubernetes.io/serviceaccount/token -login-meta=pod=k8snamespace/test-pod -tls-disabled -envoy-admin-bind-port=19000 -- --base-id 0"},
-					{"/bin/sh", "-ec", "consul-dataplane -addresses=\"1.1.1.1\" -grpc-port=8502 -proxy-service-id=$(cat /consul/connect-inject/proxyid-web-admin) " +
-						"-service-node-name=${DP_SERVICE_NODE_NAME} -log-level=info -log-json=false -envoy-concurrency=0 -credential-type=login -login-auth-method=test-auth-method " +
-						"-login-bearer-token-path=/consul/serviceaccount-web-admin/token -login-meta=pod=k8snamespace/test-pod -tls-disabled -envoy-admin-bind-port=19001 -- --base-id 1"},
+				expArgs = []string{
+					"-addresses 1.1.1.1 -grpc-port=8502 -proxy-service-id-path=/consul/connect-inject/proxyid-web " +
+						"-log-level=info -log-json=false -envoy-concurrency=0 -credential-type=login -login-auth-method=test-auth-method " +
+						"-login-bearer-token-path=/var/run/secrets/kubernetes.io/serviceaccount/token -login-meta=pod=k8snamespace/test-pod -tls-disabled -envoy-admin-bind-port=19000 -- --base-id 0",
+					"-addresses 1.1.1.1 -grpc-port=8502 -proxy-service-id-path=/consul/connect-inject/proxyid-web-admin " +
+						"-log-level=info -log-json=false -envoy-concurrency=0 -credential-type=login -login-auth-method=test-auth-method " +
+						"-login-bearer-token-path=/consul/serviceaccount-web-admin/token -login-meta=pod=k8snamespace/test-pod -tls-disabled -envoy-admin-bind-port=19001 -- --base-id 1",
 				}
 			}
 			expSAVolumeMounts := []corev1.VolumeMount{
@@ -394,10 +392,10 @@ func TestHandlerConsulDataplaneSidecar_Multiport(t *testing.T) {
 				},
 			}
 
-			for i, expCmd := range expCommand {
+			for i, expCmd := range expArgs {
 				container, err := w.consulDataplaneSidecar(testNS, pod, multiPortInfos[i])
 				require.NoError(t, err)
-				require.Equal(t, expCmd, container.Command)
+				require.Equal(t, expCmd, strings.Join(container.Args, " "))
 
 				if w.AuthMethod != "" {
 					require.Equal(t, container.VolumeMounts, []corev1.VolumeMount{
@@ -673,7 +671,7 @@ func TestHandlerConsulDataplaneSidecar_EnvoyExtraArgs(t *testing.T) {
 
 			c, err := h.consulDataplaneSidecar(testNS, *tc.pod, multiPortInfo{})
 			require.NoError(t, err)
-			require.Contains(t, c.Command[2], tc.expectedExtraArgs)
+			require.Contains(t, strings.Join(c.Args, " "), tc.expectedExtraArgs)
 		})
 	}
 }
@@ -1040,7 +1038,7 @@ func TestHandlerConsulDataplaneSidecar_Metrics(t *testing.T) {
 				require.Contains(t, err.Error(), c.expErr)
 			} else {
 				require.NoError(t, err)
-				require.Contains(t, container.Command[2], c.expCmdArgs)
+				require.Contains(t, strings.Join(container.Args, " "), c.expCmdArgs)
 			}
 		})
 	}